split roles and users endpoint
