+++ app/factory.py
@@ -29,6 +29,7 @@ from app.utils.variables import (
     ROUTER__PARSE,
     ROUTER__RERANK,
     ROUTER__SEARCH,
+    ROUTER__USERS,
 )
 
 logger = logging.getLogger(__name__)
@@ -77,6 +78,7 @@ def create_app(db_func=None, *args, **kwargs) -> FastAPI:
         roles,
         search,
         usage,
+        users,
     )
     from app.helpers._accesscontroller import AccessController
 
@@ -165,6 +167,10 @@ def create_app(db_func=None, *args, **kwargs) -> FastAPI:
     if ROUTER__USAGE not in settings.general.disabled_routers:
         add_hooks(router=usage.router)
         app.include_router(router=usage.router, tags=[ROUTER__USAGE.title()], prefix="/v1")
+
+    if ROUTER__USERS not in settings.general.disabled_routers:
+        add_hooks(router=users.router)
+        app.include_router(router=users.router, tags=[ROUTER__USERS.title()])
 
     # DEPRECATED LEGACY ENDPOINTS
     if ROUTER__COMPLETIONS not in settings.general.disabled_routers:
