+++ app/tests/integ/cassettes/test_roles.test_web_search_limits_chat_completions.yaml
